requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
discord.py==2.5.2
frozenlist
idna
multidict
propcache
python-dotenv
yarl
